backend/src/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
pytest==9.0.2
# Additional dependencies for enhanced functionality
pydantic-ai-slim==1.52.0
pydantic-ai==1.52.0
logfire==4.21.0
opik==1.9.98